refactor(mobile): split 4 god-component screens — 4,617 → 1,274 LOC orchestrators (Phase 1 #4 batch 1)
Phase 1 item #4 from `docs/planning/styrby-improve-19Apr.md`. Splits four
heavy mobile screens into orchestrator + sub-components per the
"Component-First Architecture" pattern. Continues the work of PR #80
(mobile settings.tsx → 7 sub-screens).

## Synthesis

Four parallel `general-purpose` agents in isolated worktrees, one per
file. Each agent: refactor, run typecheck + jest in worktree, report.
Reports were synthesized onto a single branch; all changes are
non-overlapping (different file trees per screen).

| File | Before → After | Sub-files | Max sub-LOC |
|------|----------------|-----------|-------------|
| `app/(tabs)/chat.tsx` | 1,185 → 398 | 10 | 337 |
| `app/(tabs)/sessions.tsx` | 1,161 → 375 | 10 | 326 |
| `app/webhooks.tsx` | 1,153 → 316 | 8 | 375 |
| `app/agent-config.tsx` | 1,118 → 185 | 17 | 307 |

**Total:** 4,617 LOC of god-components → 1,274 orchestrators + 45 focused
sub-files (every file under the 400-LOC ceiling).

## Code-reviewer pass

Independent `general-purpose` reviewer ran on the synthesized diff,
found 1 BLOCKER + 5 SUGGESTIONS + 4 NITS. This commit addresses:

- **BLOCKER (resolved)** — Five new pieces of testable logic shipped
  untested. Added 5 new test files (47 new tests):
  - `components/sessions/__tests__/grouping.test.ts` (13 tests)
  - `components/agent-config/__tests__/utils.test.ts` (12 tests)
  - `components/webhooks/__tests__/webhook-helpers.test.ts` (15 tests)
  - `hooks/__tests__/useSessionTags.test.ts` (5 tests)
  - `hooks/__tests__/useTeamMembership.test.ts` (5 tests)
- **NIT #7 (resolved)** — `agent-config/index.ts` now exports
  `useAgentConfig` (matches barrel pattern of chat/sessions/webhooks).
- **NIT #8 (resolved)** — `chat/index.ts` no longer re-exports
  `DECRYPTION_FAILED_PLACEHOLDER` (internal implementation detail of
  chat-session.ts and useRelayMessageHandler).

## Deferred to follow-up PR (test debt + architectural cleanup)

- **Hook render-tests** for `useChatSend` (186 LOC), `useRelayMessageHandler`
  (230 LOC), `useAgentConfig` (307 LOC) — need react-test-renderer +
  complex Supabase/relay mocking; separate scope.
- **SUGGESTION #2** — Inconsistent type co-location across the four
  agents (`webhooks/types` is a re-export shim; others moved sheet props
  to `types/{domain}.ts`). Decide one convention and apply.
- **SUGGESTION #3** — Some "shared" types are only used in one place
  (`SessionCardProps`, `RiskBadge`). Move back to their consumer files.
- **SUGGESTION #4** — `SessionsFilterBar.tsx` at 326 LOC has 5 near-
  identical chip rows; extract a `<FilterChip>` primitive (would drop
  to ~120 LOC).
- **SUGGESTION #5** — `useChatSend` and `useRelayMessageHandler` take
  12+ React setters as props; refactor to reducer + dispatch.
- **SUGGESTION #6** — Inconsistent path-alias usage (`../../../hooks/`
  vs `@/hooks/`). Standardize on `@/`.
- **NIT #9 + #10** — Pre-existing `eslint-disable-next-line` in
  use-agent-config.ts; silent error swallow in useTeamMembership. Both
  carried over from before the refactor.

## Verification

- `pnpm typecheck` — green
- `pnpm jest` — **941/941 tests pass** (was 886; +55 from new test files)
- `.gitignore` updated to exclude `.claude/` (worktree artifacts from
  parallel-agent workflow)

## Standards

Per `~/.claude/CLAUDE.md` "Component-First Architecture":
- Page files are orchestrators only (state, fetching, grid layout). Max 400 lines. ✅
- Every UI section gets its own component in `src/components/{page-name}/`. ✅
- Types in `src/types/{domain}.ts` not inline in page files. ✅
- Barrel exports (`index.ts`) in each component directory. ✅
- Every export carries JSDoc + WHY comments per CLAUDE.md §"Code Documentation Standards". ✅
