+++ .gitignore
@@ -126,3 +126,4 @@ ARCHITECTURE.md
 .a11y/
 .a11y-reports/
 .launch-reports/
+.claude/
