+++ packages/styrby-mobile/src/components/sessions/index.ts
@@ -1,0 +1,33 @@
+/**
+ * Sessions screen — barrel exports.
+ *
+ * WHY: Per CLAUDE.md "Component-First Architecture", every component
+ * directory exposes a single barrel so the orchestrator imports a flat
+ * surface area instead of reaching into individual files.
+ */
+
+export { SessionCard } from './SessionCard';
+export { SessionsSearchBar } from './SessionsSearchBar';
+export type { SessionsSearchBarProps } from './SessionsSearchBar';
+export { SessionsFilterBar } from './SessionsFilterBar';
+export type { SessionsFilterBarProps } from './SessionsFilterBar';
+export {
+  SessionsLoadingState,
+  SessionsErrorState,
+  SessionsEmptyState,
+} from './SessionsStateViews';
+export type {
+  SessionsErrorStateProps,
+  SessionsEmptyStateProps,
+} from './SessionsStateViews';
+export { groupSessionsByDate, formatSectionDate, getDateKey } from './grouping';
+export {
+  AGENT_CONFIG,
+  STATUS_COLORS,
+  STATUS_LABELS,
+  STATUS_CHIPS,
+  AGENT_CHIPS,
+  SCOPE_CHIPS,
+  DATE_RANGE_CHIPS,
+  getAgentConfig,
+} from './constants';
