+++ packages/styrby-mobile/src/components/webhooks/index.ts
@@ -1,0 +1,22 @@
+/**
+ * Webhooks Sub-Components Barrel
+ *
+ * Re-exports the components that compose the Webhooks orchestrator screen
+ * (`app/webhooks.tsx`). Following the project's "Component-First Architecture"
+ * pattern: each sub-component owns one cohesive concern, the orchestrator
+ * owns state and routing.
+ */
+
+export { EventBadge } from './event-badge';
+export { WebhookListItem } from './webhook-list-item';
+export { WebhookFormSheet } from './webhook-form-sheet';
+export { WebhookDetailSheet } from './webhook-detail-sheet';
+export { PowerTierGate } from './power-tier-gate';
+export {
+  EVENT_OPTIONS,
+  EVENT_COLORS,
+  WebhookFormSchema,
+  truncateUrl,
+  formatDate,
+  formatRelativeTime,
+} from './webhook-helpers';
